otorisasi admin, user dan guest

+++ foodstore-server/app/product/controller.js
@@ -3,12 +3,25 @@ const Product = require('./model');
 const Category = require('../category/model');
 const Tag = require('../tag/model');
 const config = require('../config');
+const { policyFor } = require('../policy');
 
 const fs = require('fs');
 const path = require('path');
 
 async function store(req, res, next) {
     try {
+
+        //--- cek policy ---/
+        let policy = policyFor(req.user);
+
+        if (!policy.can('create', 'Product')) {
+
+            return res.json({
+                error: 1,
+                message: `Anda tidak memiliki akses untuk membuat produk`
+            });
+
+        }
 
         let payload = req.body;
 
@@ -167,6 +180,15 @@ async function index(req, res, next) {
 
 async function update(req, res, next) {
     try {
+        let policy = policyFor(req.user);
+        if (!policy.can('update', 'Product')) {
+
+            return res.json({
+                error: 1,
+                message: `Anda tidak memiliki akses untuk membuat produk`
+            });
+
+        }
 
         let payload = req.body;
 
@@ -281,6 +303,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
     try {
+        //--- cek policy ---/
+        let policy = policyFor(req.user);
+        if (!policy.can('delete', 'Product')) { // <-- can delete
+            return res.json({
+                error: 1,
+                message: `Anda tidak memiliki akses untuk menghapus produk`
+            });
+        }
+
         let product = await Product.findOneAndDelete({ _id: req.params.id });
 
         let currentImage = `${config.rootPath}/public/upload/${product.image_url}`;
